refactor: move credential from 1.3.3 to 1.4.2
