+++ src/credential/issuance/index.ts
@@ -2,4 +2,4 @@ import * as Errors from "./common/errors";
 export { Errors };
 export type * from "./api";
 export { Issuance as V1_0_0 } from "./v1.0.0";
-export { Issuance as V1_3_3 } from "./v1.3.3";
+export { Issuance as V1_4_2 } from "./v1.4.2";
