+++ src/credential/issuance/v1.4.2/01-evaluate-issuer-trust.ts
@@ -2,7 +2,7 @@ import {
   fetchMetadata,
   type MetadataResponseV1_3,
 } from "@pagopa/io-wallet-oid4vci";
-import { sdkConfigV1_3 } from "../../../utils/config";
+import { sdkConfigV1_4 } from "../../../utils/config";
 import type { IssuanceApi } from "../api";
 import { mapToIssuerConfig } from "./mappers";
 
@@ -11,7 +11,7 @@ export const evaluateIssuerTrust: IssuanceApi["evaluateIssuerTrust"] = async (
   context = {}
 ) => {
   const issuerMetadata = (await fetchMetadata({
-    config: sdkConfigV1_3,
+    config: sdkConfigV1_4,
     credentialIssuerUrl: issuerUrl,
     authorizationServer: context.authorizationServer,
     callbacks: {
