+++ src/credential/issuance/v1.4.2/06-verify-and-parse-credential.ts
