+++ src/credential/issuance/v1.4.2/__tests__/06-verify-and-parse-credential.test.ts
