+++ src/credential/status/index.ts
@@ -1,3 +1,3 @@
 export * from "./api";
 export { CredentialStatus as V1_0_0 } from "./v1.0.0";
-export { CredentialStatus as V1_3_3 } from "./v1.3.3";
+export { CredentialStatus as V1_4_2 } from "./v1.4.2";
