+++ src/credential/trustmark/v1.4.2/get-credential-trustmark.ts
